add category filter in new_content & popular, pages are deleted when chapter & manga are deleted

+++ src/Controller/DeleteArticleController.php
@@ -20,6 +20,11 @@ final class DeleteArticleController extends AbstractController
         $articleId = $article->getId();
         $user = $this->getUser();
         $pseudo = ($user instanceof User) ? $user->getPseudo() : 'Inconnu';
+
+        if ($article->getUser() !== $user) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas supprimer l\'article d\'un autre utilisateur.');
+        }
+
         try {
             $em->remove($article);
             $em->flush();
